requirements.txt
flask==3.0.0
rembg[cpu]==2.0.53
pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
onnxruntime==1.16.3
